Moving collection service tests to collection model spec
